Existing equipment tab replicated, charInfo bugfix
Equipment tab now mirrors the original html page (with modifications to support the Characters. Work on equipment page will be delayed until it is decided what purpose this page will employ and how best to execute that. A best-in-slot calculator is still the plan, but will involve tracking both what equipment is 'available' (probably implement an inventory on the side to enable/disable items from the algo) and what is 'equipped' separately, then using the dropdown on the slots to select what is equipped (and highlighting suggestions) or perhaps find another way to indicate suggestions? Also need to write an algo or develop a simple rating system for items based on various builds, in addition to uploading json with all items and images. For now, focus will be on implementing talents. Bug was fixed where character information changes werent applied to the active char.